Pull Partition and PartitionReader classes apart
The DSV2 partition classes are made on the driver and transmitted to the
executor. This means the object needs to be serializable and weak
references aren't serializable :/ So, what we do is split everything
apart so the cache is only instantiated at the last minute on the
executor. We use a singleton pattern to make sure there is just a single
global cache.