Make a partition per-basket
We want to split a single root file into multiple partitions
to gain parallelism.

It's not the best algorithm, but for now the way I dwecide to split a
file is to split on the zeroth branch's basket boundaries. In the ideal
case, these basket boundaries are also the cluster boundaries, but
that's probably not the case. This at least gets the splitting
implemented right, and the algorithm can be tweaked later