Distinguish between in-memory and on-disk types
This is prep work to try and correctly parse unsigned integers. This
brings up an interesting problem because the on-disk and in-memory sizes
of the types are different (so it's not just the matter of simply
casting the values). The general idea is to completely implement the
disk and memory buffers separately while I sort out exactly the
different interactions I need to handle, then once things are
functional, I can pare away the extra buffer.