Support unsigned 8-bit integers
To support the full range, we upconvert the 8-bit unsigned integer to
16-bit signed integers as early as possible. This way none of the
downstream code needs to be aware of the weirdness. Need to copy-paste
the extra types (16/32 bit unsigned ints), then I can reuse the same
logic to support 64-bit unsigned ints (which upconvert to doubles) and
the root float16/double32 types.