Log all reads against "big" test data
Wire in support to log all of the application level reads against our
"large" test data. After each test-case, the cumulative reads against
each file are compactified, de-duplicated and written to a temporary
file called "unittest_ioprofile.txt". This file has a relatively simple
format and is intended to be used by a soon-to-be-written minifier
script:

Each input file is declared by a line beginning with

    #BEGIN <filename>

Then, there are N lines of the format

    <beginning offset> <ending offset>

Corresponding to bytes actually requested by the application. After each
file's reads are output, there is a final line

    #EOF

Which signifies that the file was completely written out w/o errors.