Fixup ArrayBuilder indexing
ArrayBuilder has to deal with and convert between a couple of indexing
schemes. Two of these schemes relate to arrays of offsets for
items/entries in baskets. One is indexed relative to the entire file,
while the other is indexed so the zeroth element is at the lower edge of
the partition we're deserializing. This logic was wrong. Update the
indexing to match the intention