Use JVM shutdown hook to cleanup unit tests
JUnit doesn't provide a way to call a teardown after all tests are run,
only something for a particular suite. I was handling it by having test
classes inherit from a class which added an afterClass hook, and hoping
for the best, but I found out the canonical way is to use the JVM
shutdown hook to perform these cleanup tasks.