Store ulong as long instead of double
Java and Spark don't have a native, unsigned, integral, 8 byte type, but
ROOT and C++ both do. To represent these values, there's two (not great)
options -- representing the value as a double (and losing precision) or
representing the value as a signed long (which will wrap for large
values). I'm not sure which is the right choice, but it seems like
signed long is the better of two evils, so let's do that instead.