Do not forcibly close files while they're in use
The TTree object holds a reference to the backing file, which lets us
later drill down and request the basket info in a lazy manner. However,
moving from an eager to lazy initialization for the basket info means
that we need to keep the open filehandle around longer. Remove the
try-with-resources to prevent closing the file.