Load branches lazily
Currently, we're grabbing the ObjLen/NBytes/KeyLen for each basket
eagerly, which is unneccessary - we don't need to load that information
until we're about to actually read the baskets. And, since that info is
scattered throughout the file, you pay for an I/O roundtrip for each
basket, which can be significant for a large file. Shuffle things around
to only get those values when the payload is being opened. A future
expansion could probably make it so that we load the metadata and the
payload in a single I/O op, but this is good enough for now