Reuse and mark ArrayBuilder threadpool
As reported in #63, the current partition helper makes a new threadpool
after each partition, which can cause exhaustion of availble threads if
the GC doesn't trigger quickly enough. It's also not a smart move.

Shift code to only initialize a single static thread pool and simply
change the max thread count if needed.