Shade Guava dependency
There's an annoying dependency resolution problem in certain (but not
all!) Spark installations when trying to use the Google Guava library.

Hadoop before some version used to leak its internal Guava dependency in
a way that made it so any other project that depended on Hadoop would
get the Guava dependency transitively. This is annoying because the
Guava version Hadoop needs is old and doesn't have the really
interesting features.

Because of some weird interactions with the dependency resolution
handling in Java, you could end up in the situation where both the
Hadoop-requested and Laurelin-requested Guava libraries would be
accessible to the the JVM and then it's a crap shoot to see which one
gets called when you tried to call library functions. (this,
incidentally, is how the issue skipped being noticed).

This patch "shades" the Guava dep -- meaning at packaging time, it
renamed and bundles in a private version of Guava, and rewrites the
Laurelin bytecode to look for the shaded version, resolving the conflict