Update SlimTBranch constructors
Prep work to slim up the serialized representation of SlimTBranch .. we
will eventually replace the basketEntryOffsets array with a trimmed
version represented by a list of Range<Long>. The serialization library
needs a constructor accepting the de-spacified version to initalize it
properly, so go ahead and refactor the constructor and add an extra one.

Along with the constructor changes, since our internal representation of
the basket entry offsets changes, we also modify basketEntryOffsets to
convert properly. A future update will drop the API requirement for
basketEntryOffsets