Cache DataSourceReaders
In Spark2.X, datasource objects have a very .. interesting lifetime
where they're created every time an antecedent operation needs one.
Since we do a lot of work to initialize the reader (loading all the
files, etc..), this is very unperformant

Utilize a cache where the key is the DataSourceOptions and the value is
the actual reader object. Unfortunately, since the readers themselves
are transient, we need to make these soft references so they're not
eagerly garbage collected