Pass LaurelinDSConfig all the way down the stack
Instead of extracting the various options in the driver and passing them
manually through each layer of the Spark callstack, remove the existing
code and simply produce a LaurelinDSConfig option at the very start and
pass that around. This makes it easier to add/remove configuration
options.

Note that this does add a bit (~1KB) of overhead per-partition, but it
should be only for the first partition since Java serialization memoizes
objects based on their identity.