Complete split to _v24 interface
Since the spark API is changing between 2.4 and 3.0 and will presumably
change again in the future, I need to split out the logic into a
spark-agnostic class heirarchy that does the work, and then N individual
adaptors which actually inherit from the interfaces Spark exposes. This
completes the change for _v24 (a _v30 will happen eventually).

Importantly, e.v.a.laurelin.spark_ttree is the version-agnostic logic
and e.v.a.laurelin.adaptor_v24 is the wrapper which actually implements
the spark interfaces. The adaptors are by design simplified and only
handles converting between the inputs and outputs