Add new path expansion implementation
Users are allowed to pass either directories or globs as paths, and the
desired behavior is to perform glob-expansion and recursively add .ROOT
files in directories. Add a new, better, implementation of this path
expansion function. This version accepts a list of paths, which lets us
elide many per-file, synchronous roundtrips if files are in the same
directory. This is important for cases where the file files are remote
to not incur a per-file RTT.