Use local filesystem for paths w/o a schema
The recently-added glob resolution code in 20095c99bdb caused surprising
behavior when a user passes in paths without a schema in an environment
where the hadoop configuration sets fs.defaultFS.

The previous behavior was to interpret these paths as being from the
local filesystem and load them with native java NIO libraries instead of
the hadoop-wrappeed FileSystem library. This is still the case after the
change, however the path/glob-resolution code uses the Hadoop FileSystem
code, which will pick up a configuration from the environment. If there
is a default filesystem set up in the Hadoop configuration, the
FileSystem library will assume any URL without a schema comes from that
filesystem, so instead of opening a local file, it will attempt to
resolve paths in HDFS (or whatever the filesystem is set to).

This is different than the past and arguably not the correct behavior,
so force the resolution code to use the local filesystem if the user
doesn't provide a scheme in the path.